refactor: move clockspec routines from listener to clockspec.c
