+++ winbuild/Makefile
@@ -38,6 +38,7 @@ SRCS=\
 	spawn.c \
 	opt.c \
 	cfg.c \
+	clockspec.c \
 	checksock.c \
 	fstype.c     \
 	log.c        \
